+++ astropy/coordinates/representation.py
@@ -8,7 +8,6 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
 import abc
-import functools
 import operator
 from collections import OrderedDict
 
@@ -19,6 +18,7 @@ from .angles import Angle, Longitude, Latitude
 from .distances import Distance
 from ..extern import six
 from ..utils import ShapedLikeNDArray
+from ..utils.compat import NUMPY_LT_1_12
 from ..utils.compat.numpy import broadcast_arrays
 
 __all__ = ["BaseRepresentation", "CartesianRepresentation",
@@ -52,16 +52,6 @@ class MetaBaseRepresentation(abc.ABCMeta):
             raise ValueError("Representation class {0} already defined".format(repr_name))
 
         REPRESENTATION_CLASSES[repr_name] = cls
-
-def _fstyle(precision, x):
-    fmt_str = '{0:.{precision}f}'
-    s = fmt_str.format(x, precision=precision)
-    s_trunc = s.rstrip('0')
-    if s_trunc[-1] == '.':
-        # Ensure there is one trailing 0 after a bare decimal point
-        return s_trunc + '0'
-    else:
-        return s_trunc
 
 
 @six.add_metaclass(MetaBaseRepresentation)
@@ -230,32 +220,42 @@ class BaseRepresentation(ShapedLikeNDArray):
                            for component in self.components]))
         return unitstr
 
+    def _array2string(self, prefix=''):
+        # Mimic numpy >=1.12 array2string, in which structured arrays are
+        # typeset taking into account all printoptions.
+        # TODO: in final numpy 1.12, the scalar case should work as well;
+        # see https://github.com/numpy/numpy/issues/8172
+        values = self._values
+        if NUMPY_LT_1_12 or self.isscalar:
+            # Mimic StructureFormat from numpy >=1.12 assuming float-only data.
+            from numpy.core.arrayprint import FloatFormat
+            opts = np.get_printoptions()
+            format_functions = [FloatFormat(np.atleast_1d(values[component]).ravel(),
+                                            precision=opts['precision'],
+                                            suppress_small=opts['suppress'])
+                                for component in self.components]
+
+            def fmt(x):
+                return '({})'.format(', '.join(format_function(field)
+                                               for field, format_function in
+                                               zip(x, format_functions)))
+            # Before 1.12, structures arrays were set as "numpystr",
+            # so that is the formmater we need to replace.
+            formatter = {'numpystr': fmt}
+        else:
+            fmt = repr
+            formatter = {}
+
+        return np.array2string(values, formatter=formatter, style=fmt,
+                               separator=', ', prefix=prefix)
+
     def __str__(self):
-        return '{0} {1:s}'.format(self._values, self._unitstr)
+        return '{0} {1:s}'.format(self._array2string(), self._unitstr)
 
     def __repr__(self):
         prefixstr = '    '
+        arrstr = self._array2string(prefix=prefixstr)
 
-        if self._values.shape == ():
-            v = [tuple([self._values[nm] for nm in self._values.dtype.names])]
-            v = np.array(v, dtype=self._values.dtype)
-        else:
-            v = self._values
-
-        names = self._values.dtype.names
-        precision = np.get_printoptions()['precision']
-        fstyle = functools.partial(_fstyle, precision)
-        format_val = lambda val: np.array2string(val, style=fstyle)
-        formatter = {
-            'numpystr': lambda x: '({0})'.format(
-                ', '.join(format_val(x[name]) for name in names))
-        }
-
-        arrstr = np.array2string(v, formatter=formatter,
-                                 separator=', ', prefix=prefixstr)
-
-        if self._values.shape == ():
-            arrstr = arrstr[1:-1]
 
         unitstr = ('in ' + self._unitstr) if self._unitstr else '[dimensionless]'
         return '<{0} ({1}) {2:s}\n{3}{4}>'.format(