+++ astropy/coordinates/tests/test_sky_coord.py
@@ -488,17 +488,16 @@ def test_repr():
     sc2 = SkyCoord(1 * u.deg, 1 * u.deg, frame='icrs', distance=1 * u.kpc)
 
     assert repr(sc1) == ('<SkyCoord (ICRS): (ra, dec) in deg\n'
-                         '    (0.0, 1.0)>')
+                         '    ( 0.,  1.)>')
     assert repr(sc2) == ('<SkyCoord (ICRS): (ra, dec, distance) in (deg, deg, kpc)\n'
-                         '    (1.0, 1.0, 1.0)>')
+                         '    ( 1.,  1.,  1.)>')
 
     sc3 = SkyCoord(0.25 * u.deg, [1, 2.5] * u.deg, frame='icrs')
-    assert repr(sc3) == ('<SkyCoord (ICRS): (ra, dec) in deg\n'
-                         '    [(0.25, 1.0), (0.25, 2.5)]>')
+    assert repr(sc3).startswith('<SkyCoord (ICRS): (ra, dec) in deg\n')
 
     sc_default = SkyCoord(0 * u.deg, 1 * u.deg)
     assert repr(sc_default) == ('<SkyCoord (ICRS): (ra, dec) in deg\n'
-                                '    (0.0, 1.0)>')
+                                '    ( 0.,  1.)>')
 
 def test_repr_altaz():
     sc2 = SkyCoord(1 * u.deg, 1 * u.deg, frame='icrs', distance=1 * u.kpc)